Change StringBuffer to work with uint.
Uint fits the intended usage (interop, DWORD) better than ulong. It is still far larger than
what can be expected from a simple buffer- Windows strings can't even go over short.MaxValue.

+++ src/Common/src/System/IO/PathInternal.Windows.StringBuffer.cs
@@ -31,7 +31,7 @@ namespace System.IO
         /// <summary>
         /// Gets the length of the root of the path (drive, share, etc.).
         /// </summary>
-        internal unsafe static ulong GetRootLength(StringBuffer path)
+        internal unsafe static uint GetRootLength(StringBuffer path)
         {
             Debug.Assert(path != null);
             if (path.Length == 0) return 0;
